REFACTOR: soc folder tree
