+++ soc/cores/__init__.py
@@ -1,3 +1,2 @@
-from .utils import add_rtl_sources
 from .dilithium import Dilithium
 from .crg import PetaliteCRG, PetaliteSimCRG, PowerBridge, PowerController
