+++ soc/platforms.py
@@ -1,38 +1,42 @@
-from litex.build.generic_platform import IOStandard, Pins, Subsignal
+import os
+from litex.build.generic_platform import GenericPlatform
 from litex.build.sim import SimPlatform
-import json
+
+from utils import load_io_from_json
 
 
-def load_io_from_json(json_path):
-    def int_or_str(s):
-        try:
-            return int(s)
-        except ValueError:
-            return s
+def add_rtl_sources(platform: GenericPlatform, top_level_dir_path: str):
+    # Force correct compilation order for Keccak
+    keccak_files = [
+        "keccak_pkg.sv",
+        "components/latch.sv",
+        "components/regn.sv",
+        "components/countern.sv",
+        "components/piso_buffer.sv",
+        "components/sipo_buffer.sv",
+        "components/sipo_buffer.sv",
+        "components/size_counter.sv",
+        "components/round_constant_gen.sv",
+        "components/round.sv",
+        "components/padding_gen.sv",
+        "stages/load_fsm.sv",
+        "stages/load_datapath.sv",
+        "stages/load_stage.sv",
+        "stages/permute_fsm.sv",
+        "stages/permute_datapath.sv",
+        "stages/permute_stage.sv",
+        "stages/dump_fsm.sv",
+        "stages/dump_datapath.sv",
+        "stages/dump_stage.sv",
+        "keccak.sv",
+    ]
 
-    with open(json_path, "r") as f:
-        raw_io_data = json.load(f)
+    dilithium_components_path = os.path.join(top_level_dir_path, "components/")
+    keccak_components_path = os.path.join(dilithium_components_path, "shake-sv/")
 
-    parsed_io_array = []
-    for raw_entry in raw_io_data:
-        parsed_entry = [raw_entry["name"], raw_entry["index"]]
-
-        if "subsignals" in raw_entry:
-            subsignals = [
-                Subsignal(name, Pins(int_or_str(signal["pins"])))
-                for name, signal in raw_entry["subsignals"].items()
-            ]
-            parsed_entry.extend(subsignals)
-
-        else:
-            parsed_entry.append(Pins(int_or_str(raw_entry["pins"])))
-
-        if raw_entry.get("iostandard"):
-            parsed_entry.append(IOStandard(raw_entry["iostandard"]))
-
-        parsed_io_array.append(tuple(parsed_entry))
-
-    return parsed_io_array
+    platform.add_sources(keccak_components_path, *keccak_files)
+    platform.add_source_dir(dilithium_components_path, recursive=False)
+    platform.add_source_dir(top_level_dir_path, recursive=False)
 
 
 class PetaliteSimPlatform(SimPlatform):
