+++ soc/utils/__init__.py
@@ -1,0 +1,4 @@
+from .common import *
+from .io import load_io_from_json
+from .parser import arg_parser
+from .sim import generate_gtkw_savefile
