+++ soc/utils/common.py
@@ -1,0 +1,10 @@
+from enum import StrEnum
+
+BYTE = 1
+KBYTE = 1024
+MBYTE = 1024**2
+
+
+class CommProtocol(StrEnum):
+    UART = "UART"
+    PCIE = "PCIE"
